% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_theta2_i}
\alias{C_theta2_i}
\title{calculates c_theta}
\usage{
C_theta2_i(xgrid, ygrid, tgrid, data, params, bounds)
}
\arguments{
\item{xgrid}{a vector of grid values for x}

\item{ygrid}{a vector of grid values for y}

\item{tgrid}{a t value}

\item{data}{a matrix of data}

\item{params}{a vector of parameters}

\item{bounds}{a vector of bounds for time, x, and y}
}
\value{
returns the product
}
\description{
calculates c_theta
}
\keyword{internal}
