% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_fit.R
\name{check_model_fit}
\alias{check_model_fit}
\title{check fit of estimated model on the reference data}
\usage{
check_model_fit(
  ref_data,
  Tmin = 0,
  Tmax,
  params,
  M_n,
  raster_list,
  mark_model,
  xy_bounds,
  include_comp_inds = FALSE,
  correction = "none",
  n_sim = 2500,
  verbose = TRUE
)
}
\arguments{
\item{ref_data}{a ppp object for the reference data}

\item{Tmin}{minimum value for T (defaults to 0)}

\item{Tmax}{maximum value for T}

\item{params}{vector of parameter estimates}

\item{M_n}{point to condition on}

\item{raster_list}{a list of raster objects}

\item{mark_model}{a trained mark model object}

\item{xy_bounds}{a vector of domain bounds (2 for x, 2 for y)}

\item{include_comp_inds}{`TRUE` or `FALSE` indicating whether to generate and use competition indices as covariates}

\item{correction}{type of correction to apply ("none" or "toroidal")}

\item{n_sim}{number of simulated datasets to generate}

\item{verbose}{`TRUE` or `FALSE` indicating whether to show progress}
}
\value{
a list of model fit summaries
}
\description{
check fit of estimated model on the reference data
}
