% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conditional_sum}
\alias{conditional_sum}
\title{calculates sum of values < t}
\usage{
conditional_sum(obs_t, eval_t, y)
}
\arguments{
\item{obs_t}{a vector of observed t values.}

\item{eval_t}{a t value.}

\item{y}{a vector of values.}
}
\value{
the conditional sum.
}
\description{
calculates sum of values < t
}
\keyword{internal}
