% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CondSumCppR}
\alias{CondSumCppR}
\title{calculates sum of values < t}
\usage{
CondSumCppR(obst, evalt, y)
}
\arguments{
\item{obst}{a vector of observed t}

\item{evalt}{a t value}

\item{y}{a vector of values}
}
\value{
the sum
}
\description{
calculates sum of values < t
}
