% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Ctheta2i}
\alias{Ctheta2i}
\title{calculates c_theta}
\usage{
Ctheta2i(xgrid, ygrid, tgrid, data, params)
}
\arguments{
\item{a}{vector of parameters}
}
\value{
returns the product
}
\description{
calculates c_theta
}
