% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_one_dim}
\alias{dist_one_dim}
\title{calculates distance in one dim}
\usage{
dist_one_dim(eval_t, obs_t)
}
\arguments{
\item{eval_t}{a t value.}

\item{obs_t}{a vector of t values.}
}
\value{
distance between a single t and the vector of all t values.
}
\description{
calculates distance in one dim
}
\keyword{internal}
