% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_self_correcting_parallel.R
\name{estimate_parameters_sc_parallel}
\alias{estimate_parameters_sc_parallel}
\title{Estimate parameters of the self-correcting model using log-likelihood maximization in parallel}
\usage{
estimate_parameters_sc_parallel(
  data,
  x_grid = NULL,
  y_grid = NULL,
  t_grid = NULL,
  delta_values = NULL,
  parameter_inits = NULL,
  upper_bounds = NULL,
  opt_algorithm = "NLOPT_GN_DIRECT_L",
  nloptr_options = list(maxeval = 400, xtol_rel = 1e-05, maxtime = NULL),
  verbose = TRUE,
  num_cores = parallel::detectCores() - 1,
  set_future_plan = FALSE
)
}
\arguments{
\item{data}{a data frame of locations and sizes labelled ("x", "y", "size").}

\item{x_grid}{a vector of grid values for x.}

\item{y_grid}{a vector of grid values for y.}

\item{t_grid}{a vector of grid values for t.}

\item{delta_values}{a vector of delta values.}

\item{parameter_inits}{a vector of parameter initialization values.}

\item{upper_bounds}{a vector of upper bounds for time, x, and y.}

\item{opt_algorithm}{the NLopt algorithm to use for maximization.}

\item{nloptr_options}{a list of named options for nloptr including "maxeval", "xtol_rel", and "maxtime".}

\item{verbose}{`TRUE` or `FALSE` indicating whether to show progress of optimization.}

\item{num_cores}{number of cores to use in parallel estimation.}

\item{set_future_plan}{`TRUE` or `FALSE` indicating whether to change the parallelization plan for use in the function.}
}
\value{
a list containing the optimal obtained values and the full results.
}
\description{
Estimate the parameters of the self-correcting model using nloptr given a set of delta values.
Makes use of parallel computation to allow the user to identify the optimal delta value more quickly
given available computational resources.
}
