% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{full_product}
\alias{full_product}
\title{calculates full product for one grid point}
\usage{
full_product(xgrid, ygrid, tgrid, data, params)
}
\arguments{
\item{xgrid}{a vector of grid values for x.}

\item{ygrid}{a vector of grid values for y.}

\item{tgrid}{a t value.}

\item{data}{a matrix of data.}

\item{params}{a vector of parameters.}
}
\value{
returns the product.
}
\description{
calculates full product for one grid point
}
\keyword{internal}
