% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{full_sc_lhood_fast}
\alias{full_sc_lhood_fast}
\title{calculates fast full self-correcting log-likelihood}
\usage{
full_sc_lhood_fast(xgrid, ygrid, tgrid, tobs, data, params, bounds)
}
\arguments{
\item{xgrid}{a vector of grid values for x.}

\item{ygrid}{a vector of grid values for y.}

\item{tgrid}{a vector of grid values for t.}

\item{tobs}{a vector of observed values for t.}

\item{data}{a matrix of times and locations.}

\item{params}{a vector of parameters.}

\item{bounds}{a vector of bounds for time, x, and y.}
}
\value{
evaluation of full log-likelihood.
}
\description{
calculates fast full self-correcting log-likelihood
}
\keyword{internal}
