% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_marked_point_process.R
\name{generate_mpp}
\alias{generate_mpp}
\title{Generate a marked process given locations and marks}
\usage{
generate_mpp(locations, marks = NULL, xy_bounds = NULL)
}
\arguments{
\item{locations}{a data frame of (x,y) locations with names "x" and "y".}

\item{marks}{a vector of marks.}

\item{xy_bounds}{a vector of domain bounds (2 for x, 2 for y).}
}
\value{
a ppp object with marks.
}
\description{
Creates an object of class "ppp" that represents a marked point pattern in the two-dimensional plane.
}
