% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interaction_st}
\alias{interaction_st}
\title{calculates spatio-temporal interaction}
\usage{
interaction_st(data, params)
}
\arguments{
\item{data}{a matrix of times and locations.}

\item{params}{a vector of parameters.}
}
\value{
a vector of interaction probabilities for every point.
}
\description{
calculates spatio-temporal interaction
}
\keyword{internal}
