% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interactionCpp_st}
\alias{interactionCpp_st}
\title{calculates interaction}
\usage{
interactionCpp_st(data, paramg)
}
\arguments{
\item{data}{a matrix of points}

\item{paramg}{a vector of parameters}
}
\value{
distance between a t and all t
}
\description{
calculates interaction
}
