% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helpers.R
\name{ldmppr-internal}
\alias{ldmppr-internal}
\alias{new_ldmppr_model_check}
\alias{new_ldmppr_sim}
\alias{new_ldmppr_mark_model}
\alias{new_ldmppr_fit}
\alias{preprocess_new_data}
\alias{rehydrate_xgb}
\alias{as_mark_model}
\alias{.build_sc_matrix}
\alias{.default_sc_param_bounds}
\alias{\%||\%}
\title{Internal helpers (not part of the public API)}
\usage{
new_ldmppr_model_check(
  combined_env,
  envs,
  curve_sets,
  sim_metrics = NULL,
  settings = list(),
  call = NULL
)

new_ldmppr_sim(
  process,
  mpp,
  realization,
  params,
  bounds,
  anchor_point,
  thinning,
  correction,
  include_comp_inds,
  competition_radius,
  call = NULL,
  meta = list()
)

new_ldmppr_mark_model(
  engine,
  fit_engine = NULL,
  xgb_raw = NULL,
  recipe = NULL,
  outcome = "size",
  feature_names = NULL,
  info = list()
)

new_ldmppr_fit(
  process,
  fit,
  fits = NULL,
  mapping = NULL,
  grid = NULL,
  data_summary = NULL,
  engine = "nloptr",
  call = NULL,
  timing = NULL
)

preprocess_new_data(object, new_data)

rehydrate_xgb(object)

as_mark_model(mark_model)

.build_sc_matrix(data, delta = NULL)

.default_sc_param_bounds(txy, upper_bounds)

a \%||\% b
}
\description{
These functions are used internally by ldmppr and are not intended to be
called directly by users.
}
\keyword{internal}
