% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldmppr-package.R
\docType{package}
\name{ldmppr-package}
\alias{ldmppr}
\alias{ldmppr-package}
\title{ldmppr: Estimate and Simulate from Location Dependent Marked Point Processes}
\description{
A suite of tools for estimating, assessing model fit, simulating from, and visualizing location dependent marked point processes characterized by regularity in the pattern. You provide a reference marked point process, a set of raster images containing location specific covariates, and select the estimation algorithm and type of mark model. `ldmppr` estimates the process and mark models and allows you to check the appropriateness of the model using a variety of diagnostic tools. Once a satisfactory model fit is obtained, you can simulate from the model and visualize the results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lanedrew/ldmppr}
  \item Report bugs at \url{https://github.com/lanedrew/ldmppr/issues}
}

}
\author{
\strong{Maintainer}: Lane Drew \email{lanetdrew@gmail.com} (\href{https://orcid.org/0009-0006-5427-4092}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Andee Kaplan (\href{https://orcid.org/0000-0002-2940-889X}{ORCID})
}

}
\keyword{internal}
