% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medium_example_data}
\alias{medium_example_data}
\title{Medium Example Data}
\format{
## `medium_example_data`
A data frame with 111 rows and 3 columns:
\describe{
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{size}{Size}
  ...
}
}
\source{
Real example dataset. Code to generate it can be found in `data_raw/medium_example_data.R`.
}
\usage{
data("medium_example_data")
}
\description{
A medium sized example dataset consisting of 111 observations in a (50m x 50m) square domain.
}
\details{
The dataset was generated using the Snodgrass dataset available at https://data.ess-dive.lbl.gov/view/doi:10.15485/2476543.

The full code to generate this dataset is available in the package's `data_raw` directory.
}
\keyword{datasets}
