% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Part1_1FullCpp}
\alias{Part1_1FullCpp}
\title{calculates part 1-1 full}
\usage{
Part1_1FullCpp(data, paramt)
}
\arguments{
\item{data}{a matrix of locations and times}

\item{paramt}{a vector of parameters}
}
\value{
full likelihood for part 1
}
\description{
calculates part 1-1 full
}
