% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Part1_1FullDkappaCpp}
\alias{Part1_1FullDkappaCpp}
\title{calculates part 1 full}
\usage{
Part1_1FullDkappaCpp(data, paramt)
}
\arguments{
\item{data}{a matrix of data}

\item{paramt}{a vector of parameters}
}
\value{
distance between a t and all t
}
\description{
calculates part 1 full
}
