% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Part1_2FullCpp}
\alias{Part1_2FullCpp}
\title{calculates part 1-2 full}
\usage{
Part1_2FullCpp(data, params)
}
\arguments{
\item{data}{a matrix of locations and times}

\item{params}{a vector of parameters}
}
\value{
full likelihood for part 2
}
\description{
calculates part 1-2 full
}
\keyword{internal}
