% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Part1_3FullCpp}
\alias{Part1_3FullCpp}
\title{calculates part 1-3}
\usage{
Part1_3FullCpp(xgrid, ygrid, tgrid, data, params, bounds)
}
\arguments{
\item{xgrid}{a vector of grid values for x}

\item{ygrid}{a vector of grid values for y}

\item{tgrid}{a t value}

\item{data}{a matrix of data}

\item{params}{a vector of parameters}

\item{bounds}{a vector of time, x, and y bounds}
}
\value{
full likelihood for part 3
}
\description{
calculates part 1-3
}
