% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Part1_3FullDkappaCpp}
\alias{Part1_3FullDkappaCpp}
\title{calculates part 1-3}
\usage{
Part1_3FullDkappaCpp(xgrid, ygrid, tgrid, data, params)
}
\arguments{
\item{xgrid}{a vector of grid values for x}

\item{ygrid}{a vector of grid values for y}

\item{tgrid}{a t value}

\item{data}{a matrix of data}

\item{params}{a vector of parameters}
}
\value{
distance between a t and all t
}
\description{
calculates part 1-3
}
