% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Part1_4FullCpp}
\alias{Part1_4FullCpp}
\title{calculates part 1-4}
\usage{
Part1_4FullCpp(data, paramg)
}
\arguments{
\item{data}{a matrix of locations and times}

\item{paramg}{a vector of parameters}
}
\value{
full likelihood for part 4
}
\description{
calculates part 1-4
}
