% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Part1_4FullDkappaCpp}
\alias{Part1_4FullDkappaCpp}
\title{calculates part 1-4}
\usage{
Part1_4FullDkappaCpp(data, paramg)
}
\arguments{
\item{a}{vector of parameters}
}
\value{
distance between a t and all t
}
\description{
calculates part 1-4
}
