% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Part2FullDkappaCpp}
\alias{Part2FullDkappaCpp}
\title{calculates part 2}
\usage{
Part2FullDkappaCpp(xgrid, ygrid, tgrid, data, param)
}
\arguments{
\item{a}{vector of parameters}
}
\value{
distance between a t and all t
}
\description{
calculates part 2
}
