% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{part_1_2_full}
\alias{part_1_2_full}
\title{calculates part 1-2 full}
\usage{
part_1_2_full(data, params)
}
\arguments{
\item{data}{a matrix of locations and times.}

\item{params}{a vector of parameters.}
}
\value{
full likelihood evaluation for part 2.
}
\description{
calculates part 1-2 full
}
\keyword{internal}
