% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{part_1_3_full}
\alias{part_1_3_full}
\title{calculates part 1-3}
\usage{
part_1_3_full(xgrid, ygrid, tgrid, data, params, bounds)
}
\arguments{
\item{xgrid}{a vector of grid values for x}

\item{ygrid}{a vector of grid values for y}

\item{tgrid}{a t value}

\item{data}{a matrix of times and locations}

\item{params}{a vector of parameters}

\item{bounds}{a vector of time, x, and y bounds}
}
\value{
full likelihood for part 3
}
\description{
calculates part 1-3
}
\keyword{internal}
