% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{part_1_4_full}
\alias{part_1_4_full}
\title{calculates part 1-4}
\usage{
part_1_4_full(data, params)
}
\arguments{
\item{data}{a matrix of times and locations}

\item{params}{a vector of parameters}
}
\value{
full likelihood for part 4
}
\description{
calculates part 1-4
}
\keyword{internal}
