% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{part_1_full}
\alias{part_1_full}
\title{calculates part 1 of the likelihood}
\usage{
part_1_full(xgrid, ygrid, tgrid, data, params, bounds)
}
\arguments{
\item{xgrid}{a vector of grid values for x}

\item{ygrid}{a vector of grid values for y}

\item{tgrid}{a t value}

\item{data}{a matrix of times and locations}

\item{params}{a vector of parameters}

\item{bounds}{a vector of bounds for time, x, and y}
}
\value{
first part of likelihood
}
\description{
calculates part 1 of the likelihood
}
\keyword{internal}
