% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{part_2_full}
\alias{part_2_full}
\title{calculates part 2 of the likelihood}
\usage{
part_2_full(xgrid, ygrid, tgrid, data, params, bounds)
}
\arguments{
\item{xgrid}{a vector of grid values for x.}

\item{ygrid}{a vector of grid values for y.}

\item{tgrid}{a vector of grid values for t.}

\item{data}{a matrix of times and locations.}

\item{params}{a vector of parameters.}

\item{bounds}{a vector of bounds for time, x, and y.}
}
\value{
full evaluation of second part of likelihood.
}
\description{
calculates part 2 of the likelihood
}
\keyword{internal}
