% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pdistC}
\alias{pdistC}
\title{calculates euclidean distance}
\usage{
pdistC(evalu, obsu)
}
\arguments{
\item{evalu}{a vector}

\item{obsu}{a matrix}
}
\value{
distance between a vector and each row of a matrix
}
\description{
calculates euclidean distance
}
