% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pdistCVec}
\alias{pdistCVec}
\title{calculates euclidean distance}
\usage{
pdistCVec(x, y)
}
\arguments{
\item{a}{vector y}
}
\value{
distance between the two vectors
}
\description{
calculates euclidean distance
}
