% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_refdata}
\alias{plot_refdata}
\title{Plot reference data}
\usage{
plot_refdata(ref_data)
}
\arguments{
\item{ref_data}{ppp object with marks or dataframe with (x, y, size)}
}
\value{
ggplot object of the reference data
}
\description{
Plot reference data
}
