% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_sim}
\alias{plot_sim}
\title{Plot simulated data realization}
\usage{
plot_sim(sim_data)
}
\arguments{
\item{sim_data}{ppp object with marks or dataframe with (x, y, size)}
}
\value{
ggplot object of the simulated data realization
}
\description{
Plot simulated data realization
}
