% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ldmpp.R
\name{pred_marks}
\alias{pred_marks}
\title{Predict values from the mark distribution}
\usage{
pred_marks(sim_realization, raster_list, size_model)
}
\arguments{
\item{sim_realization}{a realization of simulated values}

\item{raster_list}{list of rasters}

\item{size_model}{a predictive model}
}
\value{
a vector of predictions
}
\description{
Predict values from the mark distribution
}
