% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prodFullCpp}
\alias{prodFullCpp}
\title{calculates full product}
\usage{
prodFullCpp(xgrid, ygrid, tgrid, data, params)
}
\arguments{
\item{a}{vector of parameters}
}
\value{
returns the product
}
\description{
calculates full product
}
