% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdistC}
\alias{rdistC}
\title{calculates distance in one dim}
\usage{
rdistC(evalt, obst)
}
\arguments{
\item{evalt}{a t value}

\item{obst}{a vector of t}
}
\value{
distance between a t and all t
}
\description{
calculates distance in one dim
}
\keyword{internal}
