% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ldmpp.R
\name{Sim.s_t_int}
\alias{Sim.s_t_int}
\title{Simulate from the self-correcting model}
\usage{
Sim.s_t_int(Tmin, Tmax, par, M_n)
}
\arguments{
\item{Tmin}{minimum value for T}

\item{Tmax}{maximum value for T}

\item{par}{vector of parameter estimates}

\item{M_n}{point to condition on}
}
\value{
a list of simulated values both thinned and unthinned
}
\description{
Simulate from the self-correcting model
}
