% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ldmpp.R
\name{Sim_spatial}
\alias{Sim_spatial}
\title{Simulate the spatial component of the self-correcting model}
\usage{
Sim_spatial(M_n, par, nsim.t)
}
\arguments{
\item{M_n}{vector of (x,y)-coordinates for largest point}

\item{par}{a vector of parameters (alpha_2, beta_2)}

\item{nsim.t}{number of points to simulate}
}
\value{
a matrix of point locations in the (x,y)-plane
}
\description{
Simulate the spatial component of the self-correcting model
}
