% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_spatial_sc}
\alias{sim_spatial_sc}
\title{Simulate the spatial component of the self-correcting model (faster)}
\usage{
sim_spatial_sc(M_n, params, nsim_t, xy_bounds)
}
\arguments{
\item{M_n}{a vector of (x,y)-coordinates for anchor/first point.}

\item{params}{a vector of parameters (alpha_2, beta_2).}

\item{nsim_t}{number of points to simulate.}

\item{xy_bounds}{vector: (ax, bx, ay, by).}
}
\value{
a matrix nsim_t x 2 of point locations (x,y).
}
\description{
Simulate the spatial component of the self-correcting model (faster)
}
\keyword{internal}
