% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ldmpp.R
\name{Sim_Temporal}
\alias{Sim_Temporal}
\title{Simulate the temporal component of the self-correcting model}
\usage{
Sim_Temporal(Tmin, Tmax, par)
}
\arguments{
\item{Tmin}{minimum time value}

\item{Tmax}{maximum time value}

\item{par}{a vector of parameters (alpha_1, beta_1, gamma_1)}
}
\value{
a list of thinned and unthinned temporal samples
}
\description{
Simulate the temporal component of the self-correcting model
}
