% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_self_correcting.R
\name{simulate_sc}
\alias{simulate_sc}
\title{Simulate from the self-correcting model}
\usage{
simulate_sc(Tmin = 0, Tmax = 1, params, M_n, xy_bounds)
}
\arguments{
\item{Tmin}{minimum value for T}

\item{Tmax}{maximum value for T}

\item{params}{vector of parameter estimates}

\item{M_n}{point to condition on}

\item{xy_bounds}{vector of lower and upper bounds for the domain (2 for x, 2 for y)}
}
\value{
a list of simulated values both thinned and unthinned
}
\description{
Simulate from the self-correcting model
}
