% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_example_data}
\alias{small_example_data}
\title{Small Example Data}
\format{
## `small_example_data`
A data frame with 106 rows and 3 columns:
\describe{
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{size}{Size}
  ...
}
}
\usage{
data("small_example_data")
}
\description{
A small example dataset for testing and examples consisting of 106 observations in a 25x25 square.
}
\keyword{datasets}
