% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spat_interaction}
\alias{spat_interaction}
\title{calculates spatial interaction}
\usage{
spat_interaction(hist, newp, params)
}
\arguments{
\item{hist}{a matrix of points (x,y), n x 2.}

\item{newp}{a new point vector (x,y), length 2.}

\item{params}{a vector of parameters (alpha2, beta2).}
}
\value{
calculated probability of new point.
}
\description{
calculates spatial interaction
}
\keyword{internal}
