% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{temporal_sc}
\alias{temporal_sc}
\title{calculates temporal likelihood}
\usage{
temporal_sc(params, eval_t, obs_t)
}
\arguments{
\item{params}{a vector of parameters (alpha_1, beta_1, gamma_1)}

\item{eval_t}{a t value}

\item{obs_t}{a vector of t}
}
\value{
full temporal likelihood evaluation
}
\description{
calculates temporal likelihood
}
\keyword{internal}
