% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{temporal_sc_cpp}
\alias{temporal_sc_cpp}
\title{calculates temporal likelihood}
\usage{
temporal_sc_cpp(par, evalt, obst)
}
\arguments{
\item{par}{a vector of parameters (alpha_1, beta_1, gamma_1)}

\item{evalt}{a t value}

\item{obst}{a vector of t}
}
\value{
full temporal likelihood evaluation
}
\description{
calculates temporal likelihood
}
