% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ldmpp.R
\name{TempSpruce}
\alias{TempSpruce}
\title{Self-correcting temporal model likelihood function}
\usage{
TempSpruce(par, evalt, obst)
}
\arguments{
\item{par}{a vector of parameters (alpha_1, beta_1, gamma_1)}

\item{evalt}{the vector of times to evaluate the likelihood at}

\item{obst}{the observed time to evaluate the likelihood at}
}
\value{
the evaluated likelihood for the temporal component
}
\description{
Self-correcting temporal model likelihood function
}
