% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{thin_st_fast}
\alias{thin_st_fast}
\title{calculates acceptance for thinning mechanism during simulation}
\usage{
thin_st_fast(data, params)
}
\arguments{
\item{data}{NumericMatrix with columns (time, x, y). Assumed sorted by time ascending.}

\item{params}{NumericVector length 3: (alpha3, beta3, gamma3}
}
\value{
LogicalVector length n of whether to keep each point (true) or thin it (false).
}
\description{
calculates acceptance for thinning mechanism during simulation
}
