% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_xgboost.R
\name{train_gbm_mark_model}
\alias{train_gbm_mark_model}
\title{Train an XGBoost model for the mark distribution}
\usage{
train_gbm_mark_model(df, raster_list, save_model = FALSE, save_path)
}
\arguments{
\item{raster_list}{the list of raster objects}

\item{x}{a data.frame containing size, x, y, and age}
}
\value{
a bundled model object
}
\description{
Train an XGBoost model for the mark distribution
}
