% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_mark_model.R
\name{train_mark_model}
\alias{train_mark_model}
\title{Train an ML model for the mark distribution}
\usage{
train_mark_model(
  df,
  raster_list,
  model_type = "xgboost",
  save_model = FALSE,
  save_path
)
}
\arguments{
\item{df}{a data.frame containing named vectors size, x, y, and time}

\item{raster_list}{the list of raster objects}

\item{model_type}{the machine learning model type}

\item{save_model}{determines whether to save the generated model}

\item{save_path}{path for saving the generated model}
}
\value{
a bundled model object
}
\description{
Train an ML model for the mark distribution
}
