% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_dist}
\alias{vec_dist}
\title{calculates euclidean distance}
\usage{
vec_dist(x, y)
}
\arguments{
\item{x}{a vector of x values.}

\item{y}{a vector of y values.}
}
\value{
the distance between the two vectors.
}
\description{
calculates euclidean distance
}
\keyword{internal}
