% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_to_mat_dist}
\alias{vec_to_mat_dist}
\title{calculates euclidean distance between a vector and a matrix}
\usage{
vec_to_mat_dist(eval_u, obs_u)
}
\arguments{
\item{eval_u}{a vector}

\item{obs_u}{a matrix}
}
\value{
distance between a vector and each row of a matrix
}
\description{
calculates euclidean distance between a vector and a matrix
}
\keyword{internal}
