% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CondSumCpp}
\alias{CondSumCpp}
\title{calculates sum of values < t}
\usage{
CondSumCpp(obst, evalt, y)
}
\arguments{
\item{a}{vector of values}
}
\value{
the sum
}
\description{
calculates sum of values < t
}
