% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_self_correcting.R
\name{estimate_parameters_sc}
\alias{estimate_parameters_sc}
\title{Estimate parameters of the self-correcting model using log-likelihood optimization}
\usage{
estimate_parameters_sc(
  data,
  x_grid = NULL,
  y_grid = NULL,
  t_grid = NULL,
  parameter_inits = NULL,
  upper_bounds = NULL,
  opt_algorithm = "NLOPT_LN_SBPLX",
  nloptr_options = list(maxeval = 400, xtol_rel = 1e-05, maxtime = NULL),
  verbose = TRUE
)
}
\arguments{
\item{data}{a matrix or data frame of times and locations in the form (time, x, y).}

\item{x_grid}{a vector of grid values for x.}

\item{y_grid}{a vector of grid values for y.}

\item{t_grid}{a vector of grid values for t.}

\item{parameter_inits}{a vector of parameter initialization values.}

\item{upper_bounds}{a vector of upper bounds for time, x, and y.}

\item{opt_algorithm}{the NLopt algorithm to use for optimization.}

\item{nloptr_options}{a list of named options for nloptr including "maxeval", "xtol_rel", and "maxtime".}

\item{verbose}{`TRUE` or `FALSE` indicating whether to show progress of optimization.}
}
\value{
an nloptr object with details of the optimization including solution.
}
\description{
Estimate the parameters of the self-correcting model using nloptr given a formatted dataset.
}
\examples{
# Load the spatstat library and spruces dataset
library(spatstat)
data(spruces)

# Create the example_data data frame
example_data <- data.frame(x = spruces$x, y = spruces$y, size = spruces$marks)

# Define the grid values
x_grid <- seq(0, 56, out.length = 5)
y_grid <- seq(0, 38, out.length = 5)
t_grid <- seq(0, 1, out.length = 5)

# Define the parameter initialization values
parameter_inits <- c(5, 20, 0.025, 2, 3.5, 0.15, 2, 0.05)

# Define the upper bounds
upper_bounds <- c(1, 56, 38)

# Estimate the parameters
estimate_parameters_sc(data = example_data,
                      x_grid = x_grid,
                      y_grid = y_grid,
                      t_grid = t_grid,
                      parameter_inits = parameter_inits,
                      upper_bounds = upper_bounds,
                      opt_algorithm = "NLOPT_LN_SBPLX",
                      nloptr_options = list(maxeval = 300,
                      xtol_rel = 1e-3,
                      maxtime = NULL),
                      verbose = TRUE)


}
