% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_self_correcting.R
\name{estimate_parameters_sc}
\alias{estimate_parameters_sc}
\title{estimate parameters of self-correcting model using log-likelihood maximization}
\usage{
estimate_parameters_sc(
  xgrid,
  ygrid,
  tgrid,
  data,
  parameter_inits,
  bounds,
  opt_algorithm = "NLOPT_GN_DIRECT_L",
  nloptr_options = list(maxeval = 400, xtol_rel = 1e-05, maxtime = NULL),
  verbose = TRUE
)
}
\arguments{
\item{xgrid}{a vector of grid values for x}

\item{ygrid}{a vector of gridyvalues for y}

\item{tgrid}{a vector of grid values for t}

\item{data}{a matrix of times and locations}

\item{parameter_inits}{a vector of parameter initialization values}

\item{bounds}{a vector of bounds for time, x, and y}

\item{opt_algorithm}{the NLopt algorithm to use for maximization}

\item{nloptr_options}{a list of named options including "maxeval", "xtol_rel", and "maxtime"}

\item{verbose}{`TRUE` or `FALSE` indicating whether to show progress of optimization}
}
\value{
an nloptr object with details of the optimization including solution
}
\description{
estimate parameters of self-correcting model using log-likelihood maximization
}
