% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_self_correcting_parallel.R
\name{estimate_parameters_sc_parallel}
\alias{estimate_parameters_sc_parallel}
\title{Estimate parameters of the self-correcting model using log-likelihood maximization in parallel}
\usage{
estimate_parameters_sc_parallel(
  xgrid,
  ygrid,
  tgrid,
  data,
  delta_vals = c(0.5, 1, 1.5),
  parameter_inits,
  bounds,
  opt_algorithm = "NLOPT_GN_DIRECT_L",
  nloptr_options = list(maxeval = 400, xtol_rel = 1e-05, maxtime = NULL),
  verbose = TRUE,
  num_cores = parallel::detectCores() - 1,
  set_future_plan = FALSE
)
}
\arguments{
\item{xgrid}{a vector of grid values for x}

\item{ygrid}{a vector of grid values for y}

\item{tgrid}{a vector of grid values for t}

\item{data}{a dataframe of locations and sizes labelled (x, y, size)}

\item{delta_vals}{a vector of delta values}

\item{parameter_inits}{a vector of parameter initialization values}

\item{bounds}{a vector of bounds for time, x, and y}

\item{opt_algorithm}{the NLopt algorithm to use for maximization}

\item{nloptr_options}{a list of named options including "maxeval", "xtol_rel", and "maxtime"}

\item{verbose}{`TRUE` or `FALSE` indicating whether to show progress of optimization}

\item{num_cores}{number of cores to use in parallel estimation}

\item{set_future_plan}{`TRUE` or `FALSE` indicating whether to change the parallelization plan for use in the function}
}
\value{
a list of the optimal found values and the full results
}
\description{
Estimate the parameters of the self-correcting model using nloptr given a set of delta values.
Makes use of parallel computation to allow the user to identify the optimal delta value more quickly
given available computational resources.
}
