% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_covars}
\alias{extract_covars}
\title{Extract covariate values from a set of rasters}
\usage{
extract_covars(locations, raster_list)
}
\arguments{
\item{locations}{a data frame of (x,y) locations with names "x" and "y".}

\item{raster_list}{a list of raster objects.}
}
\value{
a matrix of covariates drawn from the provided rasters.
}
\description{
Extract covariate values from a set of rasters
}
