% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rasters.R
\name{extract_covars}
\alias{extract_covars}
\title{Extract covariate values from a set of rasters}
\usage{
extract_covars(x, raster_list)
}
\arguments{
\item{x}{a 2-column matrix or data.frame(x, y) or (lon, lat)}

\item{raster_list}{a list of raster objects}
}
\value{
a matrix of covariates drawn from the provided rasters
}
\description{
Extract covariate values from a set of rasters
}
