% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{full_sc_lhood}
\alias{full_sc_lhood}
\title{calculates negative full self-correcting log-likelihood}
\usage{
full_sc_lhood(xgrid, ygrid, tgrid, data, params, bounds)
}
\arguments{
\item{xgrid}{a vector of grid values for x}

\item{ygrid}{a vector of grid values for y}

\item{tgrid}{a vector of grid values for t}

\item{data}{a matrix of times and locations}

\item{params}{a vector of parameters}

\item{bounds}{a vector of bounds for time, x, and y}
}
\value{
negative full log-likelihood
}
\description{
calculates negative full self-correcting log-likelihood
}
