% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ldmpp.R
\name{generate_mpp}
\alias{generate_mpp}
\title{Generate a marked process given locations and marks}
\usage{
generate_mpp(location_data, marks, window)
}
\arguments{
\item{location_data}{a set of locations}

\item{marks}{a vector of marks}

\item{window}{a vector of window bounds (2 for x, 2 for y)}
}
\value{
a ppp object with marks
}
\description{
Generate a marked process given locations and marks
}
