% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interactionCpp}
\alias{interactionCpp}
\title{calculates spatial interaction}
\usage{
interactionCpp(Hist, newp, par)
}
\arguments{
\item{Hist}{a matrix of points}

\item{newp}{a new point vector}

\item{par}{a vector of parameters}
}
\value{
probability of new point
}
\description{
calculates spatial interaction
}
