% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interactionCpp}
\alias{interactionCpp}
\title{calculates interaction}
\usage{
interactionCpp(Hist, newp, par)
}
\arguments{
\item{a}{vector of parameters}
}
\value{
distance between a t and all t
}
\description{
calculates interaction
}
