% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interactionCpp_st}
\alias{interactionCpp_st}
\title{calculates spatio-temporal interaction}
\usage{
interactionCpp_st(data, paramg)
}
\arguments{
\item{data}{a matrix of times and locations}

\item{paramg}{a vector of parameters}
}
\value{
interaction probabilities for every point
}
\description{
calculates spatio-temporal interaction
}
\keyword{internal}
