% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Part2FullCpp}
\alias{Part2FullCpp}
\title{calculates part 2}
\usage{
Part2FullCpp(xgrid, ygrid, tgrid, data, param, bounds)
}
\arguments{
\item{xgrid}{a vector of grid values for x}

\item{ygrid}{a vector of grid values for y}

\item{tgrid}{a t value}

\item{data}{a matrix of data}

\item{param}{a vector of parameters}

\item{bounds}{a vector of bounds for time, x, and y}
}
\value{
distance between a t and all t
}
\description{
calculates part 2
}
