% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_marks.R
\name{predict_marks}
\alias{predict_marks}
\title{Predict values from the mark distribution}
\usage{
predict_marks(sim_realization, raster_list, size_model)
}
\arguments{
\item{sim_realization}{a realization of simulated values}

\item{raster_list}{a list of rasters}

\item{size_model}{a predictive model}
}
\value{
a vector of predicted mark values
}
\description{
Predict values from the mark distribution
}
