% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_marks.R
\name{predict_marks}
\alias{predict_marks}
\title{Predict values from the mark distribution}
\usage{
predict_marks(
  sim_realization,
  raster_list = NULL,
  mark_model = NULL,
  xy_bounds = NULL,
  include_comp_inds = FALSE,
  competition_radius = 15,
  correction = "none"
)
}
\arguments{
\item{sim_realization}{a data frame containing a thinned or unthinned realization from \code{simulate_sc}.}

\item{raster_list}{a list of raster objects.}

\item{mark_model}{a model object (typically from \code{train_mark_model}).}

\item{xy_bounds}{a vector of domain bounds (2 for x, 2 for y).}

\item{include_comp_inds}{`TRUE` or `FALSE` indicating whether to generate and use competition indices as covariates.}

\item{competition_radius}{distance for competition radius if \code{include_comp_inds} is `TRUE`.}

\item{correction}{type of correction to apply ("none" or "toroidal").}
}
\value{
a vector of predicted mark values.
}
\description{
Predict values from the mark distribution
}
