% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_marks.R
\name{predict_marks}
\alias{predict_marks}
\title{Predict values from the mark distribution}
\usage{
predict_marks(
  sim_realization,
  raster_list,
  size_model,
  xy_bounds,
  include_comp_inds = FALSE,
  correction = "none"
)
}
\arguments{
\item{sim_realization}{a realization of simulated values}

\item{raster_list}{a list of rasters}

\item{size_model}{a predictive model}

\item{xy_bounds}{a vector of domain bounds (2 for x, 2 for y)}

\item{include_comp_inds}{`TRUE` or `FALSE` indicating whether to generate and use competition indices as covariates}

\item{correction}{type of correction to apply ("none" or "toroidal")}
}
\value{
a vector of predicted mark values
}
\description{
Predict values from the mark distribution
}
