% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_spatial_sc}
\alias{sim_spatial_sc}
\title{Simulate the spatial component of the self-correcting model}
\usage{
sim_spatial_sc(M_n, params, nsim_t, xy_bounds)
}
\arguments{
\item{M_n}{a vector of (x,y)-coordinates for largest point.}

\item{params}{a vector of parameters (alpha_2, beta_2).}

\item{nsim_t}{number of points to simulate.}

\item{xy_bounds}{vector of lower and upper bounds for the domain (2 for x, 2 for y).}
}
\value{
a matrix of point locations in the (x,y)-plane.
}
\description{
Simulate the spatial component of the self-correcting model
}
\keyword{internal}
