% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_self_correcting.R
\name{sim_temporal_sc}
\alias{sim_temporal_sc}
\title{Simulate the temporal component of the self-correcting model}
\usage{
sim_temporal_sc(Tmin, Tmax, par)
}
\arguments{
\item{Tmin}{minimum time value}

\item{Tmax}{maximum time value}

\item{par}{a vector of parameters (alpha_1, beta_1, gamma_1)}
}
\value{
a list of thinned and unthinned temporal samples
}
\description{
Simulate the temporal component of the self-correcting model
}
