% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_example_data}
\alias{small_example_data}
\title{Small Example Data}
\format{
## \code{small_example_data}
A data frame with 121 rows and 3 columns:
\describe{
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{size}{Size}
  ...
}
}
\source{
Simulated dataset. Code to generate it can be found in \code{data_raw/small_example_data.R}.
}
\usage{
data("small_example_data")
}
\description{
A small example dataset for testing and examples consisting of 121 observations in a (25m x 25m) square domain.
}
\details{
The dataset was generated using the example raster data and an exponential decay size function.

The full code to generate this dataset is available in the package's \code{data_raw} directory.
}
\keyword{datasets}
