% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spat_interaction}
\alias{spat_interaction}
\title{calculates spatial interaction}
\usage{
spat_interaction(Hist, newp, params)
}
\arguments{
\item{Hist}{a matrix of points}

\item{newp}{a new point vector}

\item{params}{a vector of parameters}
}
\value{
probability of new point
}
\description{
calculates spatial interaction
}
\keyword{internal}
