% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_self_correcting.R
\name{temporal_sc}
\alias{temporal_sc}
\title{Self-correcting temporal model likelihood function}
\usage{
temporal_sc(par, evalt, obst)
}
\arguments{
\item{par}{a vector of parameters (alpha_1, beta_1, gamma_1)}

\item{evalt}{the vector of times to evaluate the likelihood at}

\item{obst}{the observed time to evaluate the likelihood at}
}
\value{
the evaluated likelihood for the temporal component
}
\description{
Self-correcting temporal model likelihood function
}
\keyword{internal}
