% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_mark_model.R
\name{train_mark_model}
\alias{train_mark_model}
\title{Train an ML model for the mark distribution}
\usage{
train_mark_model(
  data,
  raster_list = NULL,
  model_type = "xgboost",
  xy_bounds = NULL,
  save_model = FALSE,
  save_path = NULL,
  parallel = TRUE,
  include_comp_inds = FALSE,
  competition_radius = NULL,
  correction = "none",
  verbose = TRUE
)
}
\arguments{
\item{data}{a data frame containing named vectors size, x, y, and time.}

\item{raster_list}{the list of raster objects.}

\item{model_type}{the machine learning model type ("xgboost" or "random_forest").}

\item{xy_bounds}{a vector of domain bounds (2 for x, 2 for y).}

\item{save_model}{`TRUE` or `FALSE` indicating whether to save the generated model.}

\item{save_path}{path for saving the generated model.}

\item{parallel}{`TRUE` or `FALSE` indicating whether to use parallelization in model training.}

\item{include_comp_inds}{`TRUE` or `FALSE` indicating whether to generate and use competition indices as covariates.}

\item{competition_radius}{distance for competition radius if \code{include_comp_inds} is `TRUE`.}

\item{correction}{type of correction to apply ("none", "toroidal", or "truncation").}

\item{verbose}{`TRUE` or `FALSE` indicating whether to show progress of model training.}
}
\value{
a bundled model object.
}
\description{
Train an ML model for the mark distribution
}
