% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_to_mat_dist}
\alias{vec_to_mat_dist}
\title{calculates euclidean distance between a vector and a matrix}
\usage{
vec_to_mat_dist(eval_u, x_col, y_col)
}
\arguments{
\item{eval_u}{a vector of x and y coordinates.}

\item{x_col}{a vector of x coordinates.}

\item{y_col}{a vector of y coordinates.}
}
\value{
a vector of distances between a vector and each row of a matrix.
}
\description{
calculates euclidean distance between a vector and a matrix
}
\keyword{internal}
